feat: create an App component file
feat: the idea of this is that you want to list all of your pokemons in
an array together with key prop with the value of crypto randomUiid
method. you want to get data from an api by using fetchData function,
first you want to declare useEffect hook within it is the next,
you want to map your array of pokemons and keys to return fetchData
and provide it with pokemon value to it's argument so it's an array
of promises.

purpose: the purpose of an array of promises and keys here is to pass
this to Cards component to provide Card component with key and promise,
to avoid refetching data from an api when the state of this App
component changes.